Add new UN20 commands and responses
- fingerprintscanner version 2019.4.0-dev3

Former-commit-id: 2a38d52ae6d62be2ca41dbb3aa9ccee07cf5bdf2

+++ fingerprintscannermock/src/main/java/com/simprints/fingerprintscannermock/simulated/v2/SimulatedFingerV2.kt
@@ -6,7 +6,7 @@ import com.simprints.fingerprintscannermock.simulated.tools.byteArrayFromHexStri
 
 enum class SimulatedFingerV2(
     val captureFingerprintResponseCode: CaptureFingerprintResponse.ResponseCode = CaptureFingerprintResponse.ResponseCode.OK,
-    val imageQuality: Short = 80.toShort(),
+    val imageQuality: Int = 80,
     val templateBytes: ByteArray = byteArrayOf()) {
 
     PERSON_1_FINGER_1_VERSION_1_GOOD_SCAN(templateBytes = byteArrayFromHexString(
